Add files via upload
Decentralized Medical Fundraising and Donation Release System  is a blockchain-based crowdfunding platform designed to raise money for patients’ medical treatments with full transparency and security. Instead of sending donations directly to the patient, funds are held in a smart contract escrow wallet and released in stages (30%, 60%, 100%) only after verification by an escrow authority. If less than 30% of the target is raised, donors are automatically refunded. NGOs or insurers can top up funds when 80–90% of the target is achieved. This ensures fair use of donations, prevents misuse, and guarantees no contribution goes to waste.